+++ src/ts/PlotBoilerplate.ts
@@ -466,6 +466,16 @@ export class PlotBoilerplate {
         handleLine: {
           color: "rgba(180,180,180,0.5)",
           lineWidth: 1
+        },
+        pathVertex: {
+          color: "#B400FF",
+          lineWidth: 1,
+          fill: true
+        },
+        controlVertex: {
+          color: "#B8D438",
+          lineWidth: 1,
+          fill: true
         }
       },
       polygon: {
@@ -1120,48 +1130,66 @@ export class PlotBoilerplate {
         );
 
         if (this.drawConfig.drawBezierHandlePoints && this.drawConfig.drawHandlePoints) {
-          if (!d.bezierCurves[c].startPoint.attr.bezierAutoAdjust) {
-            if (d.bezierCurves[c].startPoint.attr.visible) {
-              draw.setCurrentId(`${d.uid}_h0`);
-              draw.setCurrentClassName(`${d.className}-start-handle`);
-              draw.diamondHandle(
+          if (d.bezierCurves[c].startPoint.attr.visible) {
+            const df = this.drawConfig.bezier.pathVertex.fill ? fill : draw;
+            df.setCurrentId(`${d.uid}_h0`);
+            df.setCurrentClassName(`${d.className}-start-handle`);
+            if (d.bezierCurves[c].startPoint.attr.bezierAutoAdjust) {
+              df.squareHandle(
+                d.bezierCurves[c].startPoint,
+                5,
+                this._handleColor(d.bezierCurves[c].startPoint, this.drawConfig.bezier.pathVertex.color)
+              );
+            } else {
+              df.diamondHandle(
                 d.bezierCurves[c].startPoint,
                 7,
-                this._handleColor(d.bezierCurves[c].startPoint, this.drawConfig.vertex.color)
+                this._handleColor(d.bezierCurves[c].startPoint, this.drawConfig.bezier.pathVertex.color)
               );
             }
-            d.bezierCurves[c].startPoint.attr.renderTime = renderTime;
           }
-          if (!d.bezierCurves[c].endPoint.attr.bezierAutoAdjust) {
-            if (d.bezierCurves[c].endPoint.attr.visible) {
-              draw.setCurrentId(`${d.uid}_h1`);
-              draw.setCurrentClassName(`${d.className}-end-handle`);
-              draw.diamondHandle(
+          d.bezierCurves[c].startPoint.attr.renderTime = renderTime;
+
+          if (d.bezierCurves[c].endPoint.attr.visible) {
+            const df = this.drawConfig.bezier.pathVertex.fill ? fill : draw;
+            df.setCurrentId(`${d.uid}_h0`);
+            df.setCurrentClassName(`${d.className}-start-handle`);
+            if (d.bezierCurves[c].endPoint.attr.bezierAutoAdjust) {
+              df.squareHandle(
+                d.bezierCurves[c].endPoint,
+                5,
+                this._handleColor(d.bezierCurves[c].endPoint, this.drawConfig.bezier.pathVertex.color)
+              );
+            } else {
+              df.diamondHandle(
                 d.bezierCurves[c].endPoint,
                 7,
-                this._handleColor(d.bezierCurves[c].endPoint, this.drawConfig.vertex.color)
+                this._handleColor(d.bezierCurves[c].endPoint, this.drawConfig.bezier.pathVertex.color)
               );
             }
-            d.bezierCurves[c].endPoint.attr.renderTime = renderTime;
           }
           if (d.bezierCurves[c].startControlPoint.attr.visible) {
-            draw.setCurrentId(`${d.uid}_h2`);
-            draw.setCurrentClassName(`${d.className}-start-control-handle`);
-            draw.circleHandle(
+            const df = this.drawConfig.bezier.controlVertex.fill ? fill : draw;
+            df.setCurrentId(`${d.uid}_h2`);
+            df.setCurrentClassName(`${d.className}-start-control-handle`);
+            df.circleHandle(
               d.bezierCurves[c].startControlPoint,
               3,
-              this._handleColor(d.bezierCurves[c].startControlPoint, "#008888")
+              this._handleColor(d.bezierCurves[c].startControlPoint, this.drawConfig.bezier.controlVertex.color)
             );
           }
           if (d.bezierCurves[c].endControlPoint.attr.visible) {
-            draw.setCurrentId(`${d.uid}_h3`);
-            draw.setCurrentClassName(`${d.className}-end-control-handle`);
-            draw.circleHandle(
+            const df = this.drawConfig.bezier.controlVertex.fill ? fill : draw;
+            df.setCurrentId(`${d.uid}_h3`);
+            df.setCurrentClassName(`${d.className}-end-control-handle`);
+            df.circleHandle(
               d.bezierCurves[c].endControlPoint,
               3,
-              this._handleColor(d.bezierCurves[c].endControlPoint, "#008888")
+              this._handleColor(d.bezierCurves[c].endControlPoint, this.drawConfig.bezier.controlVertex.color)
             );
           }
+          d.bezierCurves[c].startPoint.attr.renderTime = renderTime;
+          d.bezierCurves[c].endPoint.attr.renderTime = renderTime;
           d.bezierCurves[c].startControlPoint.attr.renderTime = renderTime;
           d.bezierCurves[c].endControlPoint.attr.renderTime = renderTime;
         } else {